Switch package name and codestyle to match ktfmt
The kdocformatter engine was contributed to ktfmt as the
kdoc engine: https://github.com/facebook/ktfmt/pull/358

As part of that merge, they switched the package name
of the library from kdocformatter to com.facebook.ktfmt.kdoc.

To make it easy to keep these two in sync, this CL makes
the same change here (and switches to ktfmt formatting using
the default style as well). At some point, kdocformatter
isn't needed since the same functionality is available in
ktfmt, but that isn't the case yet (e.g. the various options
for converting markup etc.)